+++ SKLibrary/src/main/java/com/silvericekey/skutilslibrary/base/BPActivity.kt
@@ -1,0 +1,5 @@
+package com.silvericekey.skutilslibrary.base
+
+open class BPActivity(var iActivity: IBaseActivity):BasePresenter() {
+
+}