+++ SKLibrary/src/main/java/com/silvericekey/skutilslibrary/base/BPFragment.kt
@@ -1,0 +1,4 @@
+package com.silvericekey.skutilslibrary.base
+
+open class BPFragment(var iFrgment: IBaseFragment) : BasePresenter() {
+}