+++ SKLibrary/src/main/java/com/silvericekey/skutilslibrary/base/BaseApplication.kt
@@ -1,9 +1,9 @@
 package com.silvericekey.skutilslibrary.base
 
 import androidx.multidex.MultiDexApplication
-import com.blankj.utilcode.util.Utils
-import com.facebook.drawee.backends.pipeline.Fresco
-import java.util.ArrayList
+import com.silvericekey.skutilslibrary.SKUtilsLibrary
+import io.realm.Realm
+import io.realm.RealmConfiguration
 
 open class BaseApplication : MultiDexApplication() {
     companion object {
@@ -12,14 +12,18 @@ open class BaseApplication : MultiDexApplication() {
 
         @JvmStatic
         fun getApp(): BaseApplication = app!!
-        var presenters: ArrayList<BasePresenter> = arrayListOf()
     }
 
     override fun onCreate() {
         super.onCreate()
         app = this
-        Utils.init(this)
-        Fresco.initialize(this)
+        SKUtilsLibrary.init(app!!)
+        Realm.init(this)
+        Realm.setDefaultConfiguration(
+                RealmConfiguration.Builder().name("realmData")
+                        .deleteRealmIfMigrationNeeded()
+                        .build()
+        )
     }
 
     open fun statusChange() {