+++ SKLibrary/src/main/java/com/silvericekey/skutilslibrary/base/IBaseFragment.kt
@@ -1,0 +1,7 @@
+package com.silvericekey.skutilslibrary.base
+
+import androidx.fragment.app.Fragment
+
+interface IBaseFragment {
+    fun getFragment():Fragment
+}