+++ app/src/main/java/com/silverknife/meizhi/features/feature_launch/LaunchPresenter.kt
@@ -1,0 +1,6 @@
+package com.silverknife.meizhi.features.feature_launch
+
+import com.silvericekey.skutilslibrary.base.BasePresenter
+
+class LaunchPresenter:BasePresenter() {
+}