+++ app/src/main/java/com/silverknife/meizhi/features/feature_test/TabsPresenter.kt
@@ -1,0 +1,10 @@
+package com.silverknife.meizhi.features.feature_test
+
+import com.silvericekey.skutilslibrary.base.BasePresenter
+
+class TabsPresenter: BasePresenter() {
+    override fun onDestroy() {
+
+    }
+
+}