+++ app/src/main/java/com/silverknife/meizhi/mvp/presenter/TabPresenter.kt
@@ -1,9 +1,0 @@
-package com.silverknife.meizhi.mvp.presenter
-
-import com.silvericekey.skutilslibrary.base.BasePresenter
-
-class TabPresenter:BasePresenter() {
-    override fun onDestroy() {
-
-    }
-}